Problem: defining structers after templates is tedious
In case when it is desirable to have multiple representations
for a fact (something that can become a fact and something
that is already a fact) there arises a lot of boilerplate code.

This requires making sure all the spelling is correct,
and defining a common trait for slots, adding a newtype for a Fact,
implementing this common trait for both, etc. etc.

Solution: implement `clips_fact` derivation via `clips_derive` tract

This derivation takes care of all that routine work.

This current implementation is a little rough on the edges, but
the basics work.