Problem: loading facts into a derived struct is impossible
The struct's field is private and can't be assigned.

Solution: implement From<Fact<'a>> for AssertedFact

This will convert Fact into an AssertedFact without
checking if the template actual matches (this can lead
to runtime errors). Should only be done if one is
absolutely sure the template is going to match.