Problem: can't pass the same Fact to different functions
However, it appears to be useful if it was possible to do
that instead of recovering that fact through search or any
other equilibristics used to obtain it in the first place.

Solution: derive Clone for Fact

This, however, carries some downsides -- if one of the clones
is retracted, other instances might be invalid. This, however,
is not a unique situation. If we obtain a Fact and then
run some arbitrary code or the rule engine and it retracts the
Fact, then we're in the same situation